+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.9.2
 - Enabled to import png by drop in the Image Editor
+- Fixed the crash bug caused by unsupported keys
+- Enabled to play the piano with mouse in the Sound Editor
 
 ## 0.9.1
 - Fixed the color pick bug of the Image Editor